chat_history_cloud_function/requirement.txt
google-cloud-datastore
flask
